Update driver to take vector double as return from FindLightSourceVect in s2.

+++ orientation_reflectance/s2.cc
@@ -59,7 +59,7 @@ main(int argc, char **argv) {
   }
 
   //retrieve the light source vectors for each sphere image
-  vector<vector<int>> direction_vectors;
+  vector<vector<double>> direction_vectors;
   vector<int> temp;
   temp = FindLightSourceVector(&an_image1,parameters[1],parameters[0],parameters[2]);
   direction_vectors.push_back(temp);
